Add a sip wrapper
When sip is executed it is passed one .sip file.  However, it uses that
to generate multiple cpp source files, one per class in the sip file.

Buildsystems need to know the outputs of commands, so this does not work
well as the output can't easily be predicted.

So, create a unity build of all files to compile into the python module.