Change command line interface
Remove the --includes flag in favor of --flags which lists all compile
flags.  That way we can additionally pass -D defines, the -fPIC flag,
the correct -std= command line option etc, meaning we can parse the
headers without errors.

Expect the client to specify the correct includes. Walking the directory
tree causes errors such as incorrect defines being set as reported
previously, such as

  /usr/include/x86_64-linux-gnu/c++/4.9.3/bits/os_defines.h:44:5: error:
function-like macro '__GLIBC_PREREQ' is not defined

 http://thread.gmane.org/gmane.comp.kde.devel.bindings/8436/focus=8446

Remove the root/header split. Require the client to specify the file to
process, and the content of the %Include to generate into the sip file.