Restore the ability to run the SIP generation tools independently of ECM.
This allows the tools to be used freely, and not least, allows for the
bulk testing of the sip_generator.py logic across a large swathe of code
(i.e. /usr/include/KF5) and thereby helps validate that the current rule
database designs are sufficiently rich for many likely coding styles.

This involved:

1. Restoring the ability for rules_engine.py to be run so as to generate
   it own documentation, and support for the following rule databases:

   - Typedef
   - Unexposed
   - MethodCode
   - TypeCode
   - ModuleCode

2. Restoring the corresponding support in sip_generator.py.

3. Restoring the sip_generation/{sip_bulk_generator.py,sip_compiler.py}
   tools in a minimally viable state.