Rule database structure changes, and related diagnostic/doc changes.
The API to the rule database is now less abstract, and more opinionated
to reduce the amount of boilerplate rule writers need to use. The
explicit ability to add new rules reflects the model of having a core
set of rules plus multiple additional sets (e.g. one per framework).

The change also improves the way that rule-based diagnostic names
are generated to match the new structure, making the diagnostic
names reflect the use of the add_rules() mechanism.